added word occurrence menu
Added the menu where the user can view the contextual examples of a selected
word and return to the main menu. This commit also packs the printing of the
words and the printing of instructions as two nested functions. This allows
for repeat printing the instructions upon exiting nested menu, as well as repeat
printing of the words whenever the user indicates with the new option "l".