added a likely name detection method
Because any individual movie or episode in the dataset usually has important,
document-specific characters, their names usually rise to the top in the TF-IDF
ranking (because of high frequency within the doc and low frequency outside it).

This commit adds an alternative word_subtitle_id function, which, in addition to
computing word frequencies in the document, also returns a dictionary of likely
names, i.e. proper nouns (the dictionary values are lists of indices of the word
in the sentences where it occurs). This will make it possible to reduce the TF-IDF
scores of likely names, and thus focus on words that are actually meaningful for a
language learner.

The name detection function errs on the side of strictness, as to maximally reduce
false positives.

Testing: added simple unit test that checks the detected likely names for the
example "detour-1945".