added functionality next/previous set of words
Previously, the user had to increase the number of displayed words in order to
see words beyond the initial list. The current functionality improves that workflow
by allowing to navigate to the next "n" or previous "p" set of N words (default 20).
The earlier option to change the number of displayed words is still available,
albeit with the "w" key (for "window").

testing: added two new test cases in test_io.py and made some minor adjustments in
the others to account for some small changes in phrasing