enabled processing source files in parallel
Previously, the code was set up to use parallelization using joblib, but it was
not really used (n=1). It has now been enabled (n=-1), and a minor adjustment was
necessary in order to avoid shared memory issues (initializing a separate instance
of the spacy language model in each of the processes called in parallel).