replace dummy code in gui
Replace the dummy code with one that actually uses function from extract words and
displays meaningful words and examples from the analyzed file. Also made a small
change to the layout separating the text box on the right into two boxes,
representing the fron and the back of a flash card.