removed spacing workaround
There is a bug with displaying whitespaces around periods and commas in the PyQt
window. However, when that text is pasted somewhere else, the whitespaces are
displayed properly, so the workaround leads to extra whitespaces in the text.

Remove the workaround, as it only has a cosmetic effect and not much practical use.