add option to save a view flashcards
Add a class to represent flashcard instances, a field to store flashcards, and a new
dialog that allows the user to view and delete any of the current flashcards.

Also includes some minor refactoring: moving all new MainWindow attribute
definitions to __init__, and replacing finding the index of a row with just
calling currentRow method.