add name filtering toggle in the GUI
The app currently uses a best-effort algorithm to try to detect names and tank
their TF-IDF scores because proper names are not useful for translation purposes.
As the results of the above algorithm are not always perfect, there is an option in
the command line version of the app to disable it. This change brings the
option to disable into the GUI as well.

testing: added a new test to test_gui.py specifically for the new functionality