temporarily remove fake German subtitle files
These were added to test the option to change the target and native language.
But they are currently messing up the TF-IDF calculations of current tests because
the new German files are treated as if they were in English.

The files should be restored after implementing a mechanism to treat source texts
in different languages separately.

testing: successfully ran test_gui.py and test_io.py, the tests affected by the
change previously.