added initial user data dialog
Added a dialog that only pops up on first usage, prompting the user for the name of
the target deck(s), and the target and native language. Currently this dialog is not
integrated into the app functionality and just prints out the stored data.

testing: added a very basic unit test for the dialog