add session selection dialog to GUI
A session represents the target and native languages and the names of the target
decks that the flash cards are written to. Currently this new dialog is just a GUI
stump that has some visual functionality: create filler sessions, delete sessions,
and proceed with the selected session. Just like the language selection dialog added
previously, the results of this dialog are not yet used anywhere in the app.

testing: the current version of the dialog is just being added for the visuals. Its
functionality will change significantly once it is integrated with the language
selection dialog, as well as the module to read user data from and write it to .json
files. A test covering all three aspects will be added then