integrated session creation menu into session selection
Renamed the LanguageSelectionDialog to SessionCreationDialog and integrated it into
the SessionSelectionDialog. Also added a one-line text edit to allow for naming
sessions. Also, reordered the classes and functions defining dialogs so that
low-level functions and classes are at the top and higher-level ones are at the
bottom.

testing: made the necessary adjustments to test_gui.py