add new function for processing directory
Add process_dir_new, which is a replacement for process_dir. The main difference
is that process_dir_new instantiates the spaCy NLP analyzer on a sentence basis,
whereas process_dir_new does so on a whole document basis. This is not only much
faster (~35s in series vs. ~82s in parallel for analyzing 10 source files), but
also leads to slightly better token-by-token assessments, thanks to the extra
context.

Removed the German-language files again in order to not complicate the new test.
Those files will be restored once the old process_dir function is removed, as its
replacement is language-aware and does not mix up results cross-language.

testing: added test_process_dir.py, which verifies that the top 10 TF-IDF results
are the same for the two process methods for the source file "detour-1945.srt".

test_detect_languages.py is failing because of the absence of the German-language
files and test_gui.py is failing because of the lack of some updates given the
addition of dynamic user sessions. Will return to those tests after the language
analysis segment is completely replaced by the more efficient and language-aware
alternative.