activate language selection in GUI
Finalize the addition of support for multiple language use cases. Now the native
and target language defined within a session are actually used to analyze the source
data files, ignore the ones in a different language, and translate to the desired
native language.

Re-added the German language data files faust_1.srt, faust_2.srt, faust_3.srt. This
is the text of Faust by Johann Wolfgang von Goethe, re-formatted as if it were the
subtitles of a 3-episode show.

testing: added new class TestTranslationGerman to the test_gui_main.py, which is
         essentially the same as TestTranslation for a German file usage.

         test_io.py and test_process_dir.py are failing when the German language
         data files are present (they path when the files are temporarily removed).
         These should be updated to not be tripped up by it, but this is exactly
         urgent.