fix split translation bug
The format of a translatable unit was "word\n\nSentence containing word." When this
unit was passed to Google Translate, the "\n\n" would be kept in the translated
output, but sometimes it wouldn't. A downstream part of the code dependend on this
format that would sometimes not be present, leading to crashes.

fix: pass to the translation thread the word and the sentence separately and do
not rely on any specific format of the translated output.

testing: added test case TestTranslationSplitBug in test_gui_main.py
         added a handy script run_tests.sh

data: added to the end of the faust_3.srt file the fake line: "Das war das größte
      Geheimdienstversagen" 4 times, in order to be able to replicate the bug
      behavior.