add session backup mechanism
Sometimes the app can crash (e.g. because of a connection timeout when connecting
to Google Translate) or it can be closed without having exported the flashcards. In
that case, any flashcards that were saved in memory but not to disk are lost.

This change adds a backup mechanism that allows writing the current flash card
to disk as soon as it is created, as well as a function to recover the data written
to disk (the session's target and native languages, the participating decks, and
the flashcards). The mechanism is not yet deployed into the GUI.

testing: added unit test test_backup_file.py, which mocks up two flashcards and some
         metadata and writes them to disk, reads them back, and verifies that the
         data is the same.